# openApiSchema.yml
openapi: 3.0.0
info:
  version: 1.0.0
  title: BlogArticles API
paths:
  /BlogArticles:
    get:
      summary: Get all BlogArticles
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BlogArticle"
    post:
      summary: Create a new BlogArticle
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BlogArticleInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogArticle"
  /BlogArticles/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
    get:
      summary: Get a BlogArticle by ID
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogArticle"
        "404":
          description: BlogArticle not found
    put:
      summary: Update a BlogArticle by ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BlogArticleInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogArticle"
        "404":
          description: BlogArticle not found
    delete:
      summary: Delete a BlogArticle by ID
      responses:
        "204":
          description: No Content
        "404":
          description: BlogArticle not found
  /Users:
    post:
      summary: Create a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /Users/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
    get:
      summary: Get a user by ID
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: User not found
    put:
      summary: Update a user by ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: User not found
    delete:
      summary: Delete a user by ID
      responses:
        "204":
          description: No Content
        "404":
          description: User not found
components:
  schemas:
    BlogArticle:
      type: object
      properties:
        id:
          type: integer
        article_id:
          type: integer
          readOnly: true
        title:
          type: string
        subtitle:
          type: string
        author:
          type: string
        date_published:
          type: string
          format: date-time
        category:
          type: string
        content:
          type: string
        meta_description:
          type: string
        reactions:
          type: object
          properties:
            likes:
              type: integer
            shares:
              type: integer
            comments:
              type: integer
        image_url:
          type: string
    BlogArticleInput:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        subtitle:
          type: string
        author:
          type: string
        category:
          type: string
        content:
          type: string
        meta_description:
          type: string
        reactions:
          type: object
          properties:
            likes:
              type: integer
            shares:
              type: integer
            comments:
              type: integer
        image_url:
          type: string
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string

    UserInput:
      type: object
      properties:
        username:
          type: string
        email:
          type: string